(refactor) Isolate utility functions from React component
